+++ query-planner-js/src/__tests__/buildPlan.test.ts
@@ -1,22 +1,32 @@
 import { QueryPlanner } from '@apollo/query-planner';
 import {
-  asFed2SubgraphDocument,
   assert,
+  asFed2SubgraphDocument,
   operationFromDocument,
   ServiceDefinition,
   Supergraph,
+  buildSubgraph,
 } from '@apollo/federation-internals';
 import gql from 'graphql-tag';
 import {
   FetchNode,
   FlattenNode,
+  PlanNode,
   SequenceNode,
+  SubscriptionNode,
   serializeQueryPlan,
 } from '../QueryPlan';
-import { FieldNode, OperationDefinitionNode, parse } from 'graphql';
+import {
+  FieldNode,
+  OperationDefinitionNode,
+  parse,
+  validate,
+  GraphQLError,
+} from 'graphql';
 import {
   composeAndCreatePlanner,
   composeAndCreatePlannerWithOptions,
+  findFetchNodes,
 } from './testHelper';
 import { enforceQueryPlannerConfigDefaults } from '../config';
 import { composeServices } from '@apollo/composition';
@@ -5042,7 +5052,156 @@ describe('Named fragments preservation', () => {
       }
     `);
   });
+
+  it('validates fragments on non-object types across spreads', () => {
+    const subgraph1 = {
+      name: 'subgraph1',
+      typeDefs: gql`
+        type Query {
+          theEntity: AnyEntity
+        }
+
+        union AnyEntity = EntityTypeA | EntityTypeB
+
+        type EntityTypeA @key(fields: "unifiedEntityId") {
+          unifiedEntityId: ID!
+          unifiedEntity: UnifiedEntity
+        }
+
+        type EntityTypeB @key(fields: "unifiedEntityId") {
+          unifiedEntityId: ID!
+          unifiedEntity: UnifiedEntity
+        }
+
+        interface UnifiedEntity {
+          id: ID!
+        }
+
+        type Generic implements UnifiedEntity @key(fields: "id") {
+          id: ID!
+        }
+
+        type Movie implements UnifiedEntity @key(fields: "id") {
+          id: ID!
+        }
+
+        type Show implements UnifiedEntity @key(fields: "id") {
+          id: ID!
+        }
+      `,
+    };
+
+    const subgraph2 = {
+      name: 'subgraph2',
+      typeDefs: gql`
+        interface Video {
+          videoId: Int!
+          taglineMessage(uiContext: String): String
+        }
+
+        interface UnifiedEntity {
+          id: ID!
+        }
+
+        type Generic implements UnifiedEntity @key(fields: "id") {
+          id: ID!
+          taglineMessage(uiContext: String): String
+        }
+
+        type Movie implements UnifiedEntity & Video @key(fields: "id") {
+          videoId: Int!
+          id: ID!
+          taglineMessage(uiContext: String): String
+        }
+
+        type Show implements UnifiedEntity & Video @key(fields: "id") {
+          videoId: Int!
+          id: ID!
+          taglineMessage(uiContext: String): String
+        }
+      `,
+    };
+
+    const [api, queryPlanner] = composeAndCreatePlannerWithOptions(
+      [subgraph1, subgraph2],
+      { reuseQueryFragments: true },
+    );
+
+    const query = gql`
+      query Search {
+        theEntity {
+          ... on EntityTypeA {
+            unifiedEntity {
+              ... on Generic {
+                taglineMessage(uiContext: "Generic")
+              }
+            }
+          }
+          ... on EntityTypeB {
+            unifiedEntity {
+              ...VideoSummary
+            }
+          }
+        }
+      }
+
+      fragment VideoSummary on Video {
+        videoId # Note: This extra field selection is necessary, so this fragment is not ignored.
+        taglineMessage(uiContext: "Video")
+      }
+    `;
+    const operation = operationFromDocument(api, query, { validate: true });
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+    const validationErrors = validateSubFetches(plan.node, subgraph2);
+    expect(validationErrors).toHaveLength(0);
+  });
 });
+
+/**
+ * For each fetch in a PlanNode validate the generated operation is actually spec valid against its subgraph schema
+ * @param plan
+ * @param subgraphs
+ */
+function validateSubFetches(
+  plan: PlanNode | SubscriptionNode | undefined,
+  subgraphDef: ServiceDefinition,
+): {
+  errors: readonly GraphQLError[];
+  serviceName: string;
+  fetchNode: FetchNode;
+}[] {
+  if (!plan) {
+    return [];
+  }
+  const fetches = findFetchNodes(subgraphDef.name, plan);
+  const results: {
+    errors: readonly GraphQLError[];
+    serviceName: string;
+    fetchNode: FetchNode;
+  }[] = [];
+  for (const fetch of fetches) {
+    const subgraphName: string = fetch.serviceName;
+    const operation: string = fetch.operation;
+    const subgraph = buildSubgraph(
+      subgraphName,
+      'http://subgraph',
+      subgraphDef.typeDefs,
+    );
+    const gql_errors = validate(
+      subgraph.schema.toGraphQLJSSchema(),
+      parse(operation),
+    );
+    if (gql_errors.length > 0) {
+      results.push({
+        errors: gql_errors,
+        serviceName: subgraphName,
+        fetchNode: fetch,
+      });
+    }
+  }
+  return results;
+}
 
 describe('Fragment autogeneration', () => {
   const subgraph = {
@@ -5058,6 +5217,7 @@ describe('Fragment autogeneration', () => {
       type A {
         x: Int
         y: Int
+        z: Int
         t: T
       }
 
@@ -5263,6 +5423,60 @@ describe('Fragment autogeneration', () => {
           fragment _generated_onA2_0 on A {
             x
             y
+          }
+        },
+      }
+    `);
+  });
+
+  it('fragments that share a hash but are not identical generate their own fragment definitions', () => {
+    const [api, queryPlanner] = composeAndCreatePlannerWithOptions([subgraph], {
+      generateQueryFragments: true,
+    });
+    const operation = operationFromDocument(
+      api,
+      gql`
+        query {
+          t {
+            ... on A {
+              x
+              y
+            }
+          }
+          t2 {
+            ... on A {
+              y
+              z
+            }
+          }
+        }
+      `,
+    );
+
+    const plan = queryPlanner.buildQueryPlan(operation);
+
+    expect(plan).toMatchInlineSnapshot(`
+      QueryPlan {
+        Fetch(service: "Subgraph1") {
+          {
+            t {
+              __typename
+              ..._generated_onA2_0
+            }
+            t2 {
+              __typename
+              ..._generated_onA2_1
+            }
+          }
+          
+          fragment _generated_onA2_0 on A {
+            x
+            y
+          }
+          
+          fragment _generated_onA2_1 on A {
+            y
+            z
           }
         },
       }
@@ -8389,6 +8603,219 @@ describe('handles fragments with directive conditions', () => {
   });
 });
 
+describe('handles operations with directives', () => {
+  const subgraphA = {
+    name: 'subgraphA',
+    typeDefs: gql`
+      directive @operation on MUTATION | QUERY | SUBSCRIPTION
+      directive @field on FIELD
+
+      type Foo @key(fields: "id") {
+        id: ID!
+        bar: String
+        t: T!
+      }
+
+      type T @key(fields: "id") {
+        id: ID!
+      }
+
+      type Query {
+        foo: Foo
+      }
+
+      type Mutation {
+        updateFoo(bar: String): Foo
+      }
+    `,
+  };
+
+  const subgraphB = {
+    name: 'subgraphB',
+    typeDefs: gql`
+      directive @operation on MUTATION | QUERY | SUBSCRIPTION
+      directive @field on FIELD
+
+      type Foo @key(fields: "id") {
+        id: ID!
+        baz: Int
+      }
+
+      type T @key(fields: "id") {
+        id: ID!
+        f1: String
+      }
+    `,
+  };
+
+  const [api, queryPlanner] = composeAndCreatePlanner(subgraphA, subgraphB);
+
+  test('if directives at the operation level are passed down to subgraph queries', () => {
+    const operation = operationFromDocument(
+      api,
+      gql`
+        query Operation @operation {
+          foo @field {
+            bar @field
+            baz @field
+            t @field {
+              f1 @field
+            }
+          }
+        }
+      `,
+    );
+
+    const queryPlan = queryPlanner.buildQueryPlan(operation);
+
+    const A_fetch_nodes = findFetchNodes(subgraphA.name, queryPlan.node);
+    expect(A_fetch_nodes).toHaveLength(1);
+    // Note: The query is expected to carry the `@operation` directive.
+    expect(parse(A_fetch_nodes[0].operation)).toMatchInlineSnapshot(`
+      query Operation__subgraphA__0 @operation {
+        foo @field {
+          __typename
+          id
+          bar @field
+          t @field {
+            __typename
+            id
+          }
+        }
+      }
+    `);
+
+    const B_fetch_nodes = findFetchNodes(subgraphB.name, queryPlan.node);
+    expect(B_fetch_nodes).toHaveLength(2);
+    // Note: The query is expected to carry the `@operation` directive.
+    expect(parse(B_fetch_nodes[0].operation)).toMatchInlineSnapshot(`
+      query Operation__subgraphB__1($representations: [_Any!]!) @operation {
+        _entities(representations: $representations) {
+          ... on Foo {
+            baz @field
+          }
+        }
+      }
+    `);
+    // Note: The query is expected to carry the `@operation` directive.
+    expect(parse(B_fetch_nodes[1].operation)).toMatchInlineSnapshot(`
+      query Operation__subgraphB__2($representations: [_Any!]!) @operation {
+        _entities(representations: $representations) {
+          ... on T {
+            f1 @field
+          }
+        }
+      }
+    `);
+  }); // end of `test`
+
+  test('if directives on mutations are passed down to subgraph queries', () => {
+    const operation = operationFromDocument(
+      api,
+      gql`
+        mutation TestMutation @operation {
+          updateFoo(bar: "something") @field {
+            id @field
+            bar @field
+          }
+        }
+      `,
+    );
+
+    const queryPlan = queryPlanner.buildQueryPlan(operation);
+
+    const A_fetch_nodes = findFetchNodes(subgraphA.name, queryPlan.node);
+    expect(A_fetch_nodes).toHaveLength(1);
+    // Note: The query is expected to carry the `@operation` directive.
+    expect(parse(A_fetch_nodes[0].operation)).toMatchInlineSnapshot(`
+      mutation TestMutation__subgraphA__0 @operation {
+        updateFoo(bar: "something") @field {
+          id @field
+          bar @field
+        }
+      }
+    `);
+  }); // end of `test`
+
+  test('if directives with arguments applied on queries are ok', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        directive @noArgs on QUERY
+        directive @withArgs(arg1: String) on QUERY
+
+        type Query {
+          test: String!
+        }
+      `,
+    };
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        directive @noArgs on QUERY
+        directive @withArgs(arg1: String) on QUERY
+      `,
+    };
+
+    const query = gql`
+      query @noArgs @withArgs(arg1: "hi") {
+        test
+      }
+    `;
+
+    const [api, qp] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const op = operationFromDocument(api, query);
+    const queryPlan = qp.buildQueryPlan(op);
+    const fetch_nodes = findFetchNodes(subgraph1.name, queryPlan.node);
+    expect(fetch_nodes).toHaveLength(1);
+    // Note: The query is expected to carry the `@noArgs` and `@withArgs` directive.
+    expect(parse(fetch_nodes[0].operation)).toMatchInlineSnapshot(`
+      query @noArgs @withArgs(arg1: "hi") {
+        test
+      }
+    `);
+  });
+
+  test('subgraph query retains the query variables used in the directives applied to the query', () => {
+    const subgraph1 = {
+      name: 'Subgraph1',
+      typeDefs: gql`
+        directive @withArgs(arg1: String) on QUERY
+
+        type Query {
+          test: String!
+        }
+      `,
+    };
+
+    const subgraph2 = {
+      name: 'Subgraph2',
+      typeDefs: gql`
+        directive @withArgs(arg1: String) on QUERY
+      `,
+    };
+
+    const query = gql`
+      query testQuery($some_var: String!) @withArgs(arg1: $some_var) {
+        test
+      }
+    `;
+
+    const [api, qp] = composeAndCreatePlanner(subgraph1, subgraph2);
+    const op = operationFromDocument(api, query);
+    const queryPlan = qp.buildQueryPlan(op);
+    const fetch_nodes = findFetchNodes(subgraph1.name, queryPlan.node);
+    expect(fetch_nodes).toHaveLength(1);
+    // Note: `($some_var: String!)` used to be missing.
+    expect(parse(fetch_nodes[0].operation)).toMatchInlineSnapshot(`
+      query testQuery__Subgraph1__0($some_var: String!) @withArgs(arg1: $some_var) {
+        test
+      }
+    `); // end of test
+  });
+}); // end of `describe`
+
 describe('@fromContext impacts on query planning', () => {
   it('fromContext variable is from same subgraph', () => {
     const subgraph1 = {
@@ -8398,13 +8825,11 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           t: T!
         }
-
         type T @key(fields: "id") @context(name: "context") {
           id: ID!
           u: U!
           prop: String!
         }
-
         type U @key(fields: "id") {
           id: ID!
           b: String!
@@ -8420,7 +8845,6 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           a: Int!
         }
-
         type U @key(fields: "id") {
           id: ID!
         }
@@ -8506,13 +8930,11 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           t: T!
         }
-
         type T @key(fields: "id") @context(name: "context") {
           id: ID!
           u: U!
           prop: String! @external
         }
-
         type U @key(fields: "id") {
           id: ID!
           field(a: String! @fromContext(field: "$context { prop }")): Int!
@@ -8527,12 +8949,10 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           a: Int!
         }
-
         type T @key(fields: "id") {
           id: ID!
           prop: String!
         }
-
         type U @key(fields: "id") {
           id: ID!
         }
@@ -8647,13 +9067,11 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           t: T!
         }
-
         type T @key(fields: "id") @context(name: "context") {
           id: ID!
           u: U!
           prop: [String]!
         }
-
         type U @key(fields: "id") {
           id: ID!
           field(a: [String]! @fromContext(field: "$context { prop }")): Int!
@@ -8668,7 +9086,6 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           a: Int!
         }
-
         type U @key(fields: "id") {
           id: ID!
         }
@@ -8719,13 +9136,11 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           t: [T]!
         }
-
         type T @key(fields: "id") @context(name: "context") {
           id: ID!
           u: U!
           prop: String!
         }
-
         type U @key(fields: "id") {
           id: ID!
           b: String!
@@ -8741,7 +9156,6 @@ describe('@fromContext impacts on query planning', () => {
         type Query {
           a: Int!
         }
-
         type U @key(fields: "id") {
           id: ID!
         }