+++ federation-integration-testsuite-js/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG for `federation-integration-testsuite-js`
 
+## 2.7.5
+
+## 2.7.4
+
+## 2.7.3
+
 ## 2.7.2
 
 ### Patch Changes