+++ query-graphs-js/CHANGELOG.md
@@ -1,5 +1,26 @@
 # CHANGELOG for `@apollo/query-graphs`
 
+## 2.7.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @apollo/federation-internals@2.7.5
+
+## 2.7.4
+
+### Patch Changes
+
+- Updated dependencies [[`d80b7f0ca1456567a0866a32d2b2abf940598f77`](https://github.com/apollographql/federation/commit/d80b7f0ca1456567a0866a32d2b2abf940598f77)]:
+  - @apollo/federation-internals@2.7.4
+
+## 2.7.3
+
+### Patch Changes
+
+- Updated dependencies [[`ec04c50b4fb832bfd281ecf9c0c2dd7656431b96`](https://github.com/apollographql/federation/commit/ec04c50b4fb832bfd281ecf9c0c2dd7656431b96), [`a494631918156f0431ceace74281c076cf1d5d51`](https://github.com/apollographql/federation/commit/a494631918156f0431ceace74281c076cf1d5d51)]:
+  - @apollo/federation-internals@2.7.3
+
 ## 2.7.2
 
 ### Patch Changes